// backend/src/types/MachineryMaintenance/maintenanceHistory.dto.ts
export type CreateMaintenanceHistoryDTO = {

    maquinariaId: number;
    fecha: string;
    costo: number;
    mecanico: string;
    descripcion: string;

}

export type UpdateMaintenanceHistoryDTO = {

    maquinariaId: number;
    fecha: string;
    costo: number;
    mecanico: string;
    descripcion: string;

}